+++ Source/NETworkManager/ViewModels/SettingsImportExportViewModel.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Windows.Input;
 using MahApps.Metro.Controls.Dialogs;
+using NETworkManager.Models.ImportExport;
 using NETworkManager.Models.Profile;
 using NETworkManager.Models.Settings;
 using NETworkManager.Utilities;
